requirements-test.txt
# CHOM API Testing Suite - Python Dependencies
# Installation: pip install -r requirements-test.txt

# Core Testing Framework
pytest>=7.4.3
pytest-html>=4.1.1
pytest-cov>=4.1.0
pytest-xdist>=3.5.0  # Parallel test execution
pytest-timeout>=2.2.0  # Test timeouts

# HTTP Client
requests>=2.31.0
urllib3>=2.1.0

# Schema Validation
jsonschema>=4.20.0
pydantic>=2.5.3

# Load Testing
locust>=2.20.0

# Test Data Generation
faker>=22.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Performance Testing
pytest-benchmark>=4.0.0

# API Documentation
openapi-spec-validator>=0.7.1

# Additional Utilities
colorama>=0.4.6
tabulate>=0.9.0
